Better banlist handling (see below)
First, banlists are now either local or central. Central banlist management
has not been implemented yet but even once it is, it can only be used by
one bot. If you're running a bot that's not called 'stretcher', you won't
be able to use this feature. You have to store your banlists locally in the
bot's directory.

Also, there are now banlists for multiple rooms. They are initialized on-
demand.